+++ CHANGELOG.md
@@ -62,6 +62,7 @@
 * pool: add `ReceiverStream` ([Yuki Kishimoto])
 * sdk: automatically resend event after NIP-42 authentication ([Yuki Kishimoto])
 * relay-builder: add NIP42 support ([Yuki Kishimoto])
+* relay-builder: add negentropy support ([Yuki Kishimoto])
 
 ### Fixed